EH structuring: try/catch/finally raising, tail-leave trimming, per-container structuring
The last big structural block. EhStructuringPass raises flat exception
regions into TryCatch/TryFinally nodes, two-phase and transactional like
StructuringPass: the whole function validates against the slice before any
mutation, and failure keeps the always-correct flat form with Regions
intact. On success Regions clear — the tree nesting is then the truth.

The slice: catch and finally handlers (filter/fault stay flat), leaves
targeting an enclosing construct's continuation, branches confined to their
region segment, endfinally closing its handler from the final block.
Handler-entry stores fold into the clause header (catch (FormatException
V_1)); the discard pop disappears; the raw rethrow prints as bare throw;.
Leaves in tail position trim to fallthrough through arbitrarily nested
constructs — the cascade lets inner bodies structure fully, so
try/catch/finally and nested-using methods print goto-free.

StructuringPass generalizes to per-container operation: every
try/catch/finally body structures (or stays flat) independently, so a
goto-heavy handler no longer flattens the rest of the method. Containers
holding a surviving leave's target stay flat to keep the label printable.
EH runs before inlining — the entry store must still be the handler's
first statement when the pass folds it.

Two real bugs fixed along the way: ldc.i4.m1 printed 65535 (ushort enum
subtraction wraps before any cast can repair it — subtract as int), and
inlining's fallthrough probe indexed function.Body.Blocks with a nested
block's ChildIndex (now resolved through the block's own container).

Corelib: 751 of 1,059 region-bearing methods (70.9%) raise structurally;
27 filter/fault excluded by design; zero crashes across the 39,468-method
sweep. Parity 43.32% -> 43.59%; remaining EH diffs are adjacent gaps
(using/lock sugar, unknown-definition truthiness), not structure.

+++ src/ILInspector.Decompiler/Pipeline/Ir/IrImporter.cs
@@ -379,10 +379,10 @@ public static class IrImporter
                     break;
 
                 case >= ILOpCode.Ldc_i4_m1 and <= ILOpCode.Ldc_i4_8:
-                    // Enum subtraction yields the enum's underlying ushort —
-                    // box an actual int or every typed-position check downstream
-                    // sees a value that lies about its own type.
-                    stack.Push(new Constant((int)(opcode - ILOpCode.Ldc_i4_0), TypeRef.CoreLib("System", "Int32")));
+                    // Subtract as int, not as the enum: ushort enum
+                    // subtraction wraps Ldc_i4_m1 - Ldc_i4_0 to 65535 before
+                    // any cast can repair it, and -1 must stay -1.
+                    stack.Push(new Constant((int)opcode - (int)ILOpCode.Ldc_i4_0, TypeRef.CoreLib("System", "Int32")));
                     break;
                 case ILOpCode.Ldc_i4_s:
                     stack.Push(new Constant((int)(sbyte)reader.ReadILByte(), TypeRef.CoreLib("System", "Int32")));
